Chore/rabbit (#150)
* Chore: code rabbit review matching (#133)

* Removed redundant index

* Created shared constants

* Updated for nullable action in domain model

* Updated for upsert

* Implemented concurrency tests

* NotImplemented on admin match endpoints

* spelling mishap in comment

* Created consistent route lint

* ensure clock injected

* Removed N+1 update

* Altered matchAction to ensure error thrown early for matches that havent been served

* The original code was marking all fetched matches as served and incrementing the batch count before filtering out matches where buildUserMatchProfile() returned null. This meant users would permanently lose
  quota for profiles that couldn't be returned (e.g., deleted user accounts).

* Fixed column equality

* Allow nulls for tests behaviour

* Updated for test clock

* Altered dodgy LocalDate

* Corrected throw on conflict of users

* Updated to use fixInstant

* Removed unnecessary transaction

* Reverted to 24 hour test

* Updated for upsertReturning

* Updated comment to reflect IllegalStateException

* Trivial Fixes

* more general fixes

* Added OpenAPI + Swagger (#148)

* Added OpenAPI + Swagger

* Updated fro code comments

* synchronized + suspend

* Exchange update

+++ common/src/main/kotlin/com/eros/common/Enums.kt
@@ -1,14 +1,16 @@
 package com.eros.common
 
+import java.math.BigDecimal
+
 /**
  * Provides the date activities along with their associated cost.
  */
-enum class DateActivity(val tokenCost: Double) {
-    DRINKS(1.0),          // Includes first drink
-    COFFEE(0.5),          // Coffee shop
-    WALK_TALK(0.5),       // Free activity, small commitment
-    DINNER(1.5),          // Meal included
-    MUSEUM(1.0),          // Entry ticket included
-    ACTIVITY(1.5),        // Bowling, mini-golf, etc.
-    BRUNCH(1.0)           // Weekend brunch
+enum class DateActivity(val tokenCost: BigDecimal) {
+    DRINKS(BigDecimal(1.0)),          // Includes first drink
+    COFFEE(BigDecimal(0.5)),          // Coffee shop
+    WALK_TALK(BigDecimal(0.5)),       // Free activity, small commitment
+    DINNER(BigDecimal(1.5)),          // Meal included
+    MUSEUM(BigDecimal(1.0)),          // Entry ticket included
+    ACTIVITY(BigDecimal(1.5)),        // Bowling, mini-golf, etc.
+    BRUNCH(BigDecimal(1.0))           // Weekend brunch
 }
